Moved toast service and combat service
