+++ resources/js/components/Tabs.vue
